Remove Almost All Default Playtime Requirements (#1660)
# Description

Yea so this codebase is supposed to be "Works out of the box", and that
doesn't apply if every antag in the game requires 100+ hours of
playtime, or that Nukies require at least 5 people readied up who each
have 200+ hours of playtime. No fucking wonder we had problems where
servers would always restart due to not being able to find any antags.

# Changelog

:cl:
- remove: Removed almost all of the EE DEFAULT playtime requirements. If
you're a downstream, set these yourself.