Arachne Silk Production Factory (#205)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Arachne can use that large abdomen of theirs to produce silk. (They
couldn't take the bionic spinnerette trait and their placed webs do not
produce silk)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] spider butt

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added ability for Arachne to produce silk

Co-authored-by: KyuPolaris <tabi@kyuudev.me>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>