Holopads, New AI Core, BSO and NTR Offices for Glacier (#1704)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description


---
Revamped the AI core in Glacier as well as added some holopads. Placed
down a NTR office, 'cuz why not.
<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/4eaeaa67-3a3f-4924-9484-66b1898af9e0)


![image](https://github.com/user-attachments/assets/ac43af89-842f-4ebf-a755-3b05eadb5ccc)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Holopads are now in Glacier.
- add: Nanotrasen Representatives and Blueshield Officers now both have
an office in Glacier. One slightly better than the other.