Fix Sprites Getting Stuck Red when Quickly Damaged (#1738)
# Description

If a damageable entity gets hit three times in a quick succession, it
will get stuck slightly red. This effect is clientside-only, but will be
present on all clients that had this happen in their PVS range.
Presumably this was happening because TryComp returns the requested
component even if it's shut down and is about to be removed.

---

# TODO

- [ ] Find a way to cope with the mental illness this gave me

---

<details><summary><h1>Media</h1></summary>
<p>

<details><summary>Before</summary>
<p>


https://github.com/user-attachments/assets/5f9a8510-8faa-49df-9cac-b9312e330f53

</p>
</details>

<details><summary>After</summary>
<p>


https://github.com/user-attachments/assets/f1d88ebb-3e6c-460d-b4cc-9e93918f1015

</p>
</details>

</p>
</details>

---

# Changelog

:cl:
- fix: Fixed damageable entities being stuck red after being damaged
several times in a quick succession.