Micro-RTG (#1744)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Before the Nanotrasen figured our microreactors, there were small
fission cells! These were, for inexplicable reasons, discontinued.
Still, the research is an essential step towards the microreactors we
know today.
The capacity of a micro-rtg is a bit lower than a microreactor, but not
by much. They take about a minute and a half to completely fill
themselves.
They also need one more sheet of uranium to make, which meant the lining
needed to be a bit thinner to still fit inside standard slots. So they
use a bit less glass and steel! Surely this has no negative
consequences.

When you're right up against a cell, it'll deal around ~0.1 radiation
damage. If you hold it, that goes up to 0.15.


---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/cdb2bdb5-a81a-487e-a8b2-8e3a68e199e7)

![image](https://github.com/user-attachments/assets/fb5272cb-f4e1-4cd6-a83e-92f255c20671)

![image](https://github.com/user-attachments/assets/d8bcf5a3-a7a6-4102-b793-16c610b62d63)


https://github.com/user-attachments/assets/dd0afce2-b613-4183-8296-8c86c4782407

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Timfa
- add: Added micro radioisotope thermoelectric generators! They're
totally safe and fun for the family!