Fixes Rejuvenation Rune to Heal Specific Damage Types Instead of Groups (#1778)
# Description

Rune of rejuvenation was trying to heal the group damage type and not
the specific damage types resulting in no healing for any brutes, cold,
or shock. Changed to the specific types of damage to actually heal. The
healing is probably a bit low especially if you get mag dumped or
laser'd but it works and can actually revive people now.

---

:cl:
- fixed: Rune of Rejuvenation to actually heal brutes, cold, and shock