Disposable Syringes (#1775)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

A cheap disposable syringe. It can draw once, inject once, and is then
rendered unusable.
The point of the syringe is to let Chemists pack specific dosages of
specific chemicals into a simple syringe. A doctor can then not
accidentally alter a dosage during treatments.

These cheap syringes live alongside normal syringes. They use fewer
materials to make, and are printed much faster.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>



https://github.com/user-attachments/assets/eb439050-b86d-49ba-b95e-22ab271f2358



https://github.com/user-attachments/assets/9e8954ec-11c9-4569-820e-08b91e09f52b

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added disposable syringes! These cheaper, quicker-to-produce
syringes let Chemists pick dosages while preparing medicine so that
doctors can inject without worrying about volume, but the syringes
cannot be used more than once.