IPCs Now Have Access to Cyber Eye Traits without Having to Pick the Initial Cyber Eyes Trait. (#1813)
Something something IPCs when gibbed have cyber eyes why do we have to
take a pointless 4 point trait

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# IPCs now have access to cyber eye traits without having to pick the
initial cyber eyes trait.

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Another mald PR, but why?
Well, simple, when you gib a IPC, you get cyber eyes, why do we have to
get the cyber eyes trait, if we already have cyber eyes, I wouldn't say
this makes IPC any more strong, if anything it makes it so they can slot
in more useful traits than a empty trait that does nothing by itself,
besides, getting rid of IPCs is still a EMP or a gun, so.


---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://i.imgur.com/Cx23LrA.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Diggy
- add: Made it so IPCs can take cyber eyes without having to waste 4
points.