Port And Update Cybersun Stealth Hardsuit (#866) (#1807)
# Description

Ports the Cybersun Stealthsuit by popular demand. I've made a few
changes to it for this codebase to better accomodate for some of our MRP
standards, including making sure that it correctly modifies the sound
emitter that all hardsuits share. Rather than deleting the sound
effects, it modifies them to be significantly quieter, with harsher
falloff such that it can only be heard to a maximum distance of 5 tiles.
Its stats and costs are rebalanced around being weaker for fighting, but
it by extension a lot cheaper than it is on LRP.

I spent more time trying to figure out how to fix the stealth movement
balance to something I was satisfied with, than I did porting the suit.

It also got a new name to match the naming scheme of the other Cybersun
suits. For EE's code, it's balanced a lot differently than it is on Goob
LRP. It's cheaper here than it is there, and more strongly favors the
"Stealth" side of balance than it does combat. If you move sufficiently
slowly, you can remain in stealth. It makes noise like a "Light"
hardsuit, but is significantly muffled. It offers less protection than
the Security tacsuits. It's really not for direct combat, and is
*strongly* meant to play to the metal gear 4 fantasy.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/db30c1a5-af8e-4748-bc52-4dcc61833eaf)

</p>
</details>

# Changelog

:cl:
- add: Ported the Cybersun Stealthsuit from Goobstation, with some
updates for the EE Codebase.

---------

Co-authored-by: starch <starchpersonal@gmail.com>