Some More BSO Options (#1816)
# Description

Adds a new "Primary Weapon" category for Blueshield Officer, which the
Chester rifle is the origin point for. Per discord discussions, this
list includes an option to replace the Chester with an energy
shield(It's way more fitting for a bodyguard to have an energy shield
oddly enough), or to pay extra points to swap it for the new BRDI-R25
rifle.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/32443925-e53d-4f00-a94b-93650211a3b4)

</p>
</details>

# Changelog

:cl:
- add: Added "Blueshield primary weapon" loadout group. His primary
selection currently consists of the classic Chester rifle, an energy
shield, and an R25.