Experimental Welder Rework (#1832)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Changes the experimental welder a decent amount. Rather than being
practically a normal welder with nigh unlimited fuel, it is now a super
fast welder with slower self-fueling, and a much lower fuel capacity.
Changes advanced industrial welder to make it still a worthwhile item.

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Advanced Industrial welder from 250 capacity to 500
- tweak: Advanced Industrial welder from 1.3x speed to 1.5x speed
- tweak: Experimental welder from 1000 capacity to 200
- tweak: Experimental welder from 1u of welder fuel generated to 0.5u
- tweak: Experimental welder from 1x speed to 2x speed

---------

Signed-off-by: EctoplasmIsGood <109397347+EctoplasmIsGood@users.noreply.github.com>
Signed-off-by: Remuchi <72476615+Remuchi@users.noreply.github.com>
Co-authored-by: Timfa <timfalken@hotmail.com>
Co-authored-by: Remuchi <72476615+Remuchi@users.noreply.github.com>