Allow Weldbot to Fix Specific Structures (#1838)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->
As requested in Goobstation MRP;

This update allows the weldbot to, _very slowly_, fix specific
structures.
To allow a structure to be repaired by a weldbot, apply the
"WeldbotFixableStructure" tag to it.
So far, this has been applied to:
- Windows
- Directional Windows
- Grounding Rods
- Tesla Coils
The robot will repair 5 damage per step to structures. In addition,
repair speeds of silicon mobs, borgs and IPC's has been lowered a bit.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Weldbots can now fix specific structures, like windows and
grounding rods.

---------

Signed-off-by: Timfa <timfalken@hotmail.com>