Literally Just the Impstation Slime Fixes (#1858)
what it says on the title.

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Got asked to port the changes to EE if I wanted to by the Impstation
folks.
Here's the link to the original PR that also showcases the changes,
markings are already in my marking PR etc.,
https://github.com/impstation/imp-station-14/pull/563

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Diggy
- tweak: Made tweaks to the slime race the same way Impstation, should
fix some transparency issues etc.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Style**
- Updated the visual appearance of a creature sprite by increasing its
opacity. The adjustment enhances the overall look by making the sprite
more defined and visually engaging. This change provides improved
clarity and vibrancy, contributing to a more polished presentation of
character design. Enjoy the refined visual experience!

<!-- end of auto-generated comment: release notes by coderabbit.ai -->