Comically Large Markings Port From Impstation + Couple Fixes (#1852)
Comically large PR where I add a bunch of Impstation marking along with
a bunch more of fixes.

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Basically what this PR does is, with permission of the people from the
great server of Impstation, it adds most of the markings they've got
there, along with a couple personal fixes that I felt were needed, which
I will explain in the changelog.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [ ] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Many markings from Impstation.
- tweak: Marking limits have been expanded, now most races should be
able to have 6 markings instead of 1 or 2 or 3.
- fix: Most races should now have access to underwear.
- remove: My sanity.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Expanded character customization with a broad range of new appearance
options including diverse markings, tattoos, scars, and hairstyles
across multiple species.
- Extended wardrobe support, now offering undershirt and underwear
options for additional species.
- Enhanced balance in visual customization by increasing point
allocations for many body parts.
- Introduced animated texture support to deliver more dynamic, fluid
visual effects for character designs.
- **New Markings**
- Added numerous new markings for various species, enhancing
customization options, including hair styles and various body markings.
- **New Localization Entries**
- Introduced new localization strings for hair accessories, markings,
and tattoos to support diverse character appearances.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: BloodfiendishOperator <141253729+Diggy0@users.noreply.github.com>