Solarian Alliance Content Part 2 (#1874)
# Description

This PR finishes up the port of Sol Alliance content from Aurora.3,
featuring EE's first new Modsuit since the Modsuits PR was first ported.
This PR mainly consists of item assets that- until they are included in
the game in some way, are mainly intended for admins to be able to run
events for and with existing antagonists.

In order to bring the new Modsuit to the game, it's been added in as a
new Sol Alliance Asset Bundle, the first of its kind for a new Faction
Kits tab available to Traitors and Nuclear Operatives. Like their
counterpart from Aurora.3, Faction Kits are intended for traitors to
roleplay as being operatives from an employer entirely different from
the standard fare. In this case, the Sol Alliance kits allow for
traitors and nuclear operatives that come from the imperialistic, human
supremacist faction known as the Sol Alliance.

# TODO

- [ ] Finish the item descriptions.

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/579eaa93-df9e-4a65-bad1-13fe2cb323ae)


![image](https://github.com/user-attachments/assets/730a8d85-7d81-48b9-82e7-970ac6eb1dca)


![image](https://github.com/user-attachments/assets/c379af22-4c49-431d-8698-33c1c47e21f1)

</p>
</details>

# Changelog

:cl:
- add: Added all remaining assets for the Sol Alliance Armed Forces, for
use in admin-events, or to be added to the game at a later time.
- add: Added a new Sol Alliance Modsuit, the CSA-85x "Mazu" Modsuit. 
- add: Added new Faction Kits tab for Traitor & Nuclear Operative
uplinks. For now, it contains kits for operatives belonging to the Sol
Alliance. Given that the Sol Alliance are Human supremacists, traitor
kits for them are restricted to Humans only.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a dedicated faction catalog featuring detailed asset
bundles for individual and team purchases related to the Solarian
Alliance.
- Expanded character customization with numerous new uniforms, tactical
gear, modsuits, headwear, and outerwear.
- Added new weapon options including distinct sabres and matching
sheaths.
- Enhanced visual assets with new texture metadata for various helmets
and clothing items, improving in-game presentation.
- Added localization entries for weapon and clothing descriptions,
enhancing contextual information for users.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->