Spray Nozzle Tweaks (#1883)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

The spray nozzle in it's current state is abysmal for any cleaning work.
This is because the projectile speed was so fast that any reagents could
barely touch any tiles. I adjusted the projectile speed so that it can
now (MOSTLY*) reach all tiles. Fire rate and other parameters were
changed. Now you can be your very own touhou boss!

* footsteps and slidemarks impact, but i don't know how to fix that
right now
---


<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>



https://github.com/user-attachments/assets/66f4d324-c880-4a6a-92b6-09f093701a8f



</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Spray Nozzle parameters have been tweaked for MAXIMUM
cleanliness.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced weapon performance with a significantly increased firing
rate.
- Introduced additional projectile settings, including adjustable speed
and a broader firing angle range, offering a more dynamic and effective
experience during gameplay.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>