Document Printer Port (#1875)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Port the document printer from corvax, full credit to the creators, I
only did the port. (Sorry for anything, first project of this scope.)
Prs:
https://github.com/space-syndicate/space-station-14/pull/1832
https://github.com/space-syndicate/space-station-14/pull/1859
https://github.com/space-syndicate/space-station-14/pull/2750
https://github.com/space-syndicate/space-station-14/pull/2957

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Captura de tela 2025-03-03
223754](https://github.com/user-attachments/assets/d4e0c1ce-54a8-492f-bf56-261bac6c019a)
![Captura de tela 2025-03-03
223842](https://github.com/user-attachments/assets/c7f16797-d57c-45dd-b250-0c492296a7f0)
![Captura de tela 2025-03-03
223908](https://github.com/user-attachments/assets/5b88a71a-5979-47e4-974f-eb92b47c521a)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Add document printer



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced comprehensive localization for lathe categories and
document types, enhancing labels for reports, applications, and
certificates.
- Added a new document printer machine that supports various document
templates and improved printing functionality.
- Expanded in-game inventory and production systems with new items,
materials, and recipes to streamline document generation and management.
- Added new entries for document types and categories to enhance
bureaucratic processes within the game.

- **Bug Fixes**
  - Resolved issues related to document printing and categorization.

- **Documentation**
- Updated localization files to reflect new entries and descriptions for
improved user experience.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Ed <96445749+TheShuEd@users.noreply.github.com>
Co-authored-by: SonicHDC <100022571+SonicHDC@users.noreply.github.com>
Co-authored-by: Morb0 <14136326+Morb0@users.noreply.github.com>
Co-authored-by: lzk <124214523+lzk228@users.noreply.github.com>
Co-authored-by: Krosus777 <38509947+Krosus777@users.noreply.github.com>
Co-authored-by: spaghetti <frikadeliSpaghetti@outlook.com>
Co-authored-by: Zekins <136648667+Zekins3366@users.noreply.github.com>