Civilian Disabler (#1882)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Spess is a dangerous pless. A weak self defence weapon is a must-have in
certain regions.
This weakened disabler will help the crew defend against burglars, giant
spiders and other critters. It deals 0.3 heat damage, and 2/3rd of the
stamina damage of a real disabler. It's also legal to carry! Though, of
course, not legal to use on people without good reason.

The idea behind the weapon is that it's basically old Disabler tech
that's very outdated and underpowered compared to modern disablers,
stuck in a shiny modern shell to make it visually appealing to the
average spessman.

---



<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


https://github.com/user-attachments/assets/7394a166-39c0-4d8f-8471-60610eff98d0


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added a civilian disabler! It deals next to no heat damage, and
only two-third of the stamina damage of a sec disabler


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
- Introduced a new civilian self-defense weapon, the "civilian
disabler," with a lower-powered option.
- Added a modified projectile variant, the "disabler bolt," featuring
refined sound, damage, and visual effects.
- Expanded crafting options by incorporating new manufacturing recipes
for the civilian disabler.
- Provided a new loadout option, "LoadoutCivilianDisabler," with
tailored job restrictions.
- Enabled access to the civilian disabler through updated research
unlocks and enhanced visual states for improved gameplay aesthetics.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: Timfa <timfalken@hotmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>