BSO Gear Tweaks (#1885)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Small tweaks to various equipment belonging to the BSO, to even things
out.

---


# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Added 50% heat resist to BSO hardsuit (unsure why it didn't
have it)
- tweak: Slightly changed the energy shotguns recharge properties to
avoid a bug and give it less downtime
- tweak: Buffed the fire rate/use time of the chester, to hopefully
allow it to keep up with other weapons again.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Enhanced battery performance on weapons with increased charge capacity
and faster recharge mechanics.
- Revamped security hardsuit design with a refined description and added
heat protection for improved defense.
- Improved weapon responsiveness by increasing firing speed and reducing
action delay.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: EctoplasmIsGood <109397347+EctoplasmIsGood@users.noreply.github.com>