Mining Shuttle & Dignitary Stuff for Arena (#1893)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

<p>
Adds the mining dock & mining shuttle console for travel to lavaland.
Adds NTR office with the NTR locker mapped. Also adds the BSO locker to
bridge without an office cause BSO is not worthy.
Map render under images. NTR office is south east of brig. Just has a
desk and a bed. No fax as the NTR already has their own fax, I feel like
having a second fax would increase where the NTR would need to look for
faxes.
Referring to the shuttle that takes one to lavaland as mining shuttle,
not the pathfinder.
</p>


<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>



![Arena](https://github.com/user-attachments/assets/a9719209-becd-4315-8b57-e0179b8eb8d7)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
no cl no fun