Terminus (#1881)
# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Added a new weapon, the DT-8 "Terminus"
It's a large melee weapon capable of firing a spread of disabler rounds
on alternate fire.
It's an arsenal tech research, and uniquely can only be wielded by onis.

(I'm being bribed to make this)
Also, I had to move the OniComponent.cs file to shared. I'm not sure if
that borked anything but tell me if it does.

(PS. Thank you VMSolidus for helping me make this ten times less
shitcoded.)

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Textures complete
- [x] Add lathe recipe
- [x] Add to research

---

<details><summary><h1>Media</h1></summary>
<p>


!![image](https://github.com/user-attachments/assets/cf02d64a-53a8-4454-9312-010fa92cf597)

</p>
</details>

---

# Changelog

:cl: SolsticeOfTheWinter, OtherPersonPlaceHolder
- add: Added a new weapon for oni security players, the terminus.

---------

Signed-off-by: Solstice <solsticeofthewinter@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>