Add RPD and Pipe Stacking (#1857)
# Summary

To add the RPD into the game for both atmos and CE's to use. Introduce
pipe stacking as well since the RPD will introduce better atmos
configuration. Most of the code was ported over from goob LRP. With some
of it being touched up on and adding a bit of quality of life to the
RPD.

This was a PR originally made for goob-mrp but requested to be moved to
EE. Check out the detailed commits here:
https://github.com/Goob-Station/Goob-Station-MRP/pull/96

---

# Description

## RPD
From the original PR made on goob-station the RPD has been ported over.
It is similar to the RCD where you will be able to make
pipes/vents/pumps/mixers quickly. Each dial on the main radial will
bring up specific pipes like ie pipes-> corner, t-shape -cross shape.

The RPD starts out in the CE and Atmos lockers with 50 charges. Each
construction and deconstruction cost 1 charge. There is also one extra
compressed matter in the starting lockers for both CE and Atmos for the
RPD.

You can research the recipe of the RPD with the reverse engineering
machine. The recipe cost of the RPD is a bit more compared to the RCD.


![RPD_UI](https://github.com/user-attachments/assets/ed2fe76b-009a-4081-abe4-7c431f21ab27)

## Pipe Stacking

With the introduction of the RPD, it was fair to check for pipes that
overlapped when making new pipes. But with introducing the code from LRP
goob station it was decided to bring over the pipe stacking feature too.
You can make certain components pipe stackable or allow if you can put
multiple of the same pipe on the same tile. If you want you change a
pipe to be non-stackable by using the component type:
PipeRestrictOverlap

Here's an example of pipe stacking. You can make a simple scrub loob
tighter or keep it normal. The freedom is given.
![pipe-stacking
example](https://github.com/user-attachments/assets/69aa19d4-2ff4-4cee-a2d1-b6709884c6e0)

Here is an extreme example. This compact design creates a ton of frezon.
Obviously, replace the trit miner with an actual burn chamber.

![image](https://github.com/user-attachments/assets/4db7b627-0d3a-4c78-b5a1-8d52551b3127)

## Belt overlay

I decided to add a simple belt overlay sprite to the game for the RPD. 

Here's what it looks like outside the belt and inside the belt

![RPD](https://github.com/user-attachments/assets/d599dbed-56d0-42c7-acee-20fa23f04f2c)

## RPD Borg Module

I decided to add experimental RPD that will self-charge. The only way to
get this RPD is through admeme spawn or being a borg. The starting
charge of the Borg module is 25 and will self-charge at the same rate at
10 seconds per charge. I also added an admeme version that charges at 1
second per charge.

Here is what the borg module looks like
![Borg
module](https://github.com/user-attachments/assets/6fd1bcff-fdb7-49ad-a8f5-e4bab86bd8f4)

The borg module can be researched in the t2 industrial branch as the
same research for the RCD module.

![Research-rpd](https://github.com/user-attachments/assets/d325dd3f-eed9-4744-9003-3b7d51387432)

---

# credit
Original RPD PR: https://github.com/Goob-Station/Goob-Station/pull/1203
Original Pipe Stacking PRs:
https://github.com/Goob-Station/Goob-Station/pull/391 
https://github.com/Goob-Station/Goob-Station/pull/389 
Prevent stacking pipes (#28308) - Original Link is dead.


---

# TODO
- [X] Add RPD
- [X] Add RPD to reverse engineering
- [x] Add toolbelt equipped sprites
- [x] Add Pipe Stacking
- [x] Add an experimental RPD module for engie borgs.

---

# Changelog

:cl:
- add: Added RPD, glory to atmosia!
- tweak: RPD now spawns in Atmos and CE lockers
- tweak: Extra compressed matter has been added into both Atmos/CE
lockers for more RPD/RCD uses
- add: You can research RPD with the reverse engineering machine.
- add: A belt overlay sprite to the CE and utility toolbelt.
- add: Pipe Stacking ported over from LRP Goob station. Compress those
frezon factories!
- add: Toolbelt overlay sprite for the RPD.
- add: RPD Borg module researched in t-2 industrial advanced tools.

---------

Co-authored-by: Nathaniel Adams <60526456+Nathaniel-Adams@users.noreply.github.com>