Embeddable Projectile Fixes (Courtesy Of White Dream) (#1907)
* embed children

* locale

* fix

* Update Resources/Locale/ru-RU/weapons/throwing/throwing.ftl



* // WWDP MY BELOVED

# Description

This is a port (and general cleanup of)
https://github.com/WWhiteDreamProject/wwdpublic/pull/278
Plus clearing a bunch of compiler warning causing things, and also a fix
for one of our Heisentests. This PR mainly fixes issues with embeddable
projectiles being potentially unintentionally destroyed when the thing
they are embedded in gets destroyed. Which can be disastrous for example
if my energy sword embeds into a wall, and also destroys that wall,
which destroys my precious energy sword...

# Changelog

:cl: Spatison
- fix: Embeddeable projectiles are no longer always destroyed when the
thing they are embedded in gets destroyed. Energy Sword throwing
enthusiasts may rejoice now that their precious sword is safe.

Co-authored-by: vanx <61917534+Vaaankas@users.noreply.github.com>
Co-authored-by: vanx <discord@vanxxxx>
Co-authored-by: Spatison <137375981+Spatison@users.noreply.github.com>