Magma Hand (#1927)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Sprited by Greenwall, YML'd by myself. The Magma Hand is a slime-like
creature that you can encounter in Lavaland. It's very hostile, but you
_can_ pet it!

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Add to Lavaland spawning list

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>
Its blood is sulfur!


![image](https://github.com/user-attachments/assets/7008c190-5502-4e7e-800a-a1027e2ef9f8)

Its attacks burn you!


https://github.com/user-attachments/assets/0bf21207-1dc2-4cf5-9b9b-89a558488963

Ice makes it die very quickly!


https://github.com/user-attachments/assets/4c742c6e-ac5e-462f-9f1e-e05ce0fcaa0b


![image](https://github.com/user-attachments/assets/feb78377-179c-47bd-991a-b88e37af7c95)


![image](https://github.com/user-attachments/assets/92a875cd-fdb2-4edf-843b-a26c671fbd99)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Greenwall
- add: Magma Hand! Sprite by Greenwall, YML by Timfa. Bring a water gun
filled with ice, and you'll be fine!