Cyber Eye Changes (#1876)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Removes the flare shielding trait, instead adding welding arc protection
to the diagnostics hud, and giving the medical hud a built in chem
scanner.

No trait will give you access to immunity to flashes however, and you
will instead be required to have proper eye protection on again to
counteract it. This makes securities lives a lot better, and makes the
flash a viable tool again.

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- remove: Removes flare shielding
- tweak: Adds eye protection to diagnostics hud
- tweak: Adds chemical analyser to med hud
- tweak: Increases med hud and diagnostics hud costs to 2


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Gameplay Balance**
- Reduced the negative impact for the `FlareShielding` trait, offering a
milder penalty.
- Removed the flash immunity effect from the `FlareShielding` trait to
streamline its behavior while retaining it in the
`FlareShieldingModule`.
- Updated requirements for the `Blindness`, `Nearsighted`, and
`Photophobia` traits by removing the dependency on the
`FlareShieldingModule`.
- Adjusted point values for several traits, reflecting a more
significant negative impact for others.
- Introduced new components, enhancing the functionality of various
traits.
- **Feature Enhancements**
- Expanded the description of the `CyberEyesMedical` trait to include a
Chemical Analysis Hud.
- Updated the `CyberEyesDiagnostic` trait description to incorporate
flare shielding for improved eye protection.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: EctoplasmIsGood <109397347+EctoplasmIsGood@users.noreply.github.com>