Fix Oni (#1935)
# Description

I didn't notice during the last PR that someone added an override for
the Tags component of Oni, otherwise I'd have said something about it.
You can't add a tag component to a species because it overrides a
massive shitton of tags that are needed for player characters to be able
to do certain things, such as DoorBumpOpener...

# Changelog

:cl:
- fix: Fixed Oni not being able to do a ton of things that required
hardcoded tags. I fucking hate the tag system.