Cyberiad 1.0: Service Rework (#1938)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Changes Box to Cyberiad, with the service area receiving a rework, in
the future there will be more updates in the depts, with the next one
probably being the Sec.
Make some fixes on Gax and Bagel.

I changed the maintainer of some maps from "Estação Pirata" to "Ichai".
I am not part of the Estação Pirata

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/e18d2441-8c01-49b9-9c6a-9d2c2c4672b6)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Ichai
- add: Cyberiad added to the rotation.
- fix: Gax: Fixed missing window in CE office, Fixed walls in SM to
prevent radiation leaks.
- tweak: Port changes to Bagel from GoobStation
- remove: Box is no more.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>