Novu-Nederic Language (#1952)
beste land

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Once the language of seafarers and merchants, now the voice of spacers,
terraformers, and pioneers. Spoken in the New-Netherlands, this direct
and pragmatic trade language carries the sharp consonants and clipped
vowels of its old-world roots.
Built on centuries of seafaring, land reclamation, and commerce, it
thrives among those who shape entire worlds — just as their ancestors
shaped the land from the sea.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Bitterbal eten

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Novu-Nederic language (incl. trait)