Cyberiad 1.1: Eng Rework (#1974)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Complete rework in engineering and AI core, eng now has a new area for
singu/tesla

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] map review by some maintainer

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<summary><h1>Media</h1></summary>


AI Core:

![image](https://github.com/user-attachments/assets/ff00912a-2ad4-47fd-81ab-e85307ebf2b5)

Engineering:

![image](https://github.com/user-attachments/assets/96aabe20-1a6e-4d4e-a4e6-54f3a6e017b2)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Ichai
- tweak: Cyberiad has had its engineering and AI core completely
reworked, enjoy the new department