Buff the Healing Properties of Cable Coils when Used on IPCs. (#1934)
# Description

Ointment heals 10 heat/cold/shock and 5 caustic, and mesh does 15 for
all of them.
Then you have cables, which while faster to apply (delay of 0.6 versus
2), heal 1.66 heat and shock. Thats it.
While yes, the stack size for cables is doubled, that still just sucks
in terms of healing per timeframe, healing per stack, etc.

While I could have gone with a buff while keeping them all the same, I
decided it may be more interesting to give each cable type a unique
use-case.

LV cables keep the delay of 0.6, but now heal 2 heat/shock and 0.5 cold.
This brings the healing up to 60 per stack. (As opposed to 150 for
ointment or 255 for mesh) The time it takes to apply a full stack is
thus still 18, as opposed to 30 for the topicals. This makes them best
for brief healing mid-combat, or for scenarios where you find a downed
IPC and need to find cables to heal them, as LVs are by far the most
common.

MV cables double everything. Double healing, and double the time taken.
So 120 per stack, 36 to apply. This means that they are strictly better
then LVs due to more healing per stack, except in situations where you
are mid combat and may not have enough time to stand still for the MV
delay, in which case you will want to use LVs since those still have the
same healing per time.

HV Cables are again double the effectiveness of MVs, bringing you up to
240 per stack, which is almost as good as mesh.
The delay has been set to 1.8, meaning these are outright better, in
healing per time, and healing per stack, but this high delay means it is
not viable in combat, and as such relegated to scenarios such as a
medbay/robo bay, or field healing by a BSO or Corpsman after all
hostiles are down.

I hope this gives each cable more of a unique use, and it would be
interesting to see a BSO or Corpsman, especially if they themselves are
an IPC, carry a combat medkit with just 2 of each cable stack inside,
which IMO is actually a fair use of the combat medkits since these
cables are just topicals but for IPCs.


Lastly, these cables all now heal some cold damage. While IPCs dont take
it due to for example spacing, what happens if you take it due to a
weapon? Things like the cultist void blade, which if thats not in-game
yet, it will be ported eventually, deal cold damage, and right now,
unless you have the nano-repair trait, that damage is permanent and
unhealable.
This change should also fix that. The cold damage healed by cables is
25% of the heat and shock damage healed by that cable type, but given
that IPCs also only TAKE 20% cold damage, that should be sufficient.

---

# Changelog

:cl: BramvanZijp
- tweak: Cable coils are now better at healing burns for IPCs, with MV
and HV cables providing more healing but being slower to apply.

---------

Signed-off-by: BramvanZijp <56019239+BramvanZijp@users.noreply.github.com>