Rename the Medsecglasses to Blueshield Glasses and Add a Diagnostic Hud to Them. (#1924)
# Description

The medsecglasses have been renamed to blueshield glasses, with an
updated description.
They also now act as a diagnostic hud, so the status of borgs and IPCs
can be monitored.


Firstly, the name looks quite odd next to for example the corpsman
glasses, especially if you want to try to fit the word diagnostic in
there too, and as these glasses are only used by the BSO, and have the
sprite to match, naming them as such seems better.

As for why the diagnostic ability is being added at all, the blueshield
is supposed to protect command from threats, and both command AND
threats could be IPCs, and borgs can also be threats if their laws are
compromised. Being able to identify their integrity at a glance to
decide your priorities, both for who to heal/shield and who to shoot,
should be a nice QOL improvement.

---

# Changelog

:cl: BramvanZijp
- tweak: The medsecglasses have been renamed to blueshield glasses,
alongside an updated description
- add: Added a diagnostic HUD to the blueshield glasses.

---------

Signed-off-by: BramvanZijp <56019239+BramvanZijp@users.noreply.github.com>