Tritium and Ammonia Canisters Resprites (#1986)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Added new sprites to replace the old tritium and ammonia canisters to be
consistent with the rest of the canister sprites.

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added new sprites for the Tritium and Ammonia canisters

Signed-off-by: Rosycup <178287475+Rosycup@users.noreply.github.com>