Power Cell Draw Fixes (#1989)
# Description

Power cell draw wasn't tick invariant, so batteries on this codebase
were being drained anywhere from 30 to 60x faster than people expect,
depending on server performance. This was notably something that
severely effected MODSuits, which can now enjoy a battery lifespan of
around 30 minutes. But it's also a big deal for IPCs.

Closes https://github.com/Simple-Station/Einstein-Engines/issues/1956

# Changelog

:cl:
- fix: Batteries no longer drain power 30 to 60 times faster than
physics states. This is particularly a large buff to IPCs and MODSuits,
who both no longer have to deal with batteries lasting an absurdly short
amount of time.
- tweak: Solarian Modsuit now provides protection from cold.