The Accursed Abomination Grows Ever Stronger (#1969)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds a bunch of random ass changes to Smile the slime; its description
says its been through thousands of experiments so I thought I'd make it
a very wacky creature. Smile can no longer do damage, but rather injects
a TINY amount of omnizine on hit, allowing it to heal people (slowly),
Smile is now functionally immortal, cannot be slowed, is unaffected by
gravity, can go under doors, and bleeds Nothing (the reagent)

Smile is basically a spectator now, besides the fact that it can heal
(and I guess take a bullet or two(hundred) for a member of the crew, if
you want to be cinematic)


----
:cl:
- tweak: Smile no longer does damage, but rather injects a tiny amount
of omnizine on hit
- tweak: Smile is no longer effected by most slowness sources
- tweak: Smile has ascended to godhood
- tweak: Smile is no longer effected by gravity
- tweak: Smile now bleeds Nothing
- tweak: Smile can now go under doors

---------

Co-authored-by: Your Name <EctoplasmIsGood@users.noreply.github.com>