Survival Medipens Inject with a Delay Based on Environment (#1954)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Miners have medipens that can take you down in two clicks with no delays
leading to an almost instant death (example in media), so this updates
miner medipens to how they work in ss13 - injections are instant under
50kpa, otherwise with a small delay

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->
Test locally:
- [x] normal pressure - surv medipens injections delayed
- [x] no pressure - surv medipens dont have delay
- [x] lavaland (~44kpa) - surv medipens dont have delay
- [x] other medipens don't have delays

Other:
- [ ] maybe split this into UnderpressurizedDelay component or
something, since more stuff from lavaland can be low-pressure locked

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very l


https://github.com/user-attachments/assets/3da673b7-4dc1-4f2d-9f97-1b6794e9b441

arge at times, so this is a good way to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


https://github.com/user-attachments/assets/bf7e393a-206e-4654-ae5b-d01fa9d1eda6

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: survival medipen works instantly only in underpressurized
(under 50kpa) environment

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>