Made Envirosuit Visors Translucent (#1960)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

What's the point of having a badass flaming skeleton if you can't
actually see the badass flaming skeleton? I made the visors of all
envirosuits translucent with a purple tint (because there should be
plasma within, though probably not enough to make it nearly opaque). I
may revisit this later to see if I would be capable of adjusting the
transparency based on the plasma levels of a connected tank, as per the
suggestion of Alia Ash/timfa.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![envirosuit
front](https://github.com/user-attachments/assets/48ed6fd8-082a-455a-80ea-c9b81b887718)

![envirosuit
side](https://github.com/user-attachments/assets/48e465e9-eab0-4ccb-94ee-5abe3905870e)

https://www.youtube.com/watch?v=nqgUG_JVzCs


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Plasmaman envirosuits have translucent visors.