Port My Merged DS14 Fork Commits (#2038)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Port the two merged PRs I did on DS14 to the main EE Repo:
https://github.com/SS14-Classic/deep-station-14/pull/133
https://github.com/SS14-Classic/deep-station-14/pull/136

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Tile variantization PR
- [x] Decal PR
- [x] Make sure this builds properly because I seem to be getting errors
on it??

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Zeta_Null
- tweak: Tweaked several tiles to make variants
- add: Added a "Worn steel tile" variant for mappers to include, just a
steel tile with a missing panel
- add: Added a couple of text decals