Add Engi Text Decal (#2052)
# Description

Quick PR to add a missing Engineering decal to go with the set of decals
I ported from DS14. (I had originally missed it when doing my PR there)

---

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/9d9c1be2-1e92-4a4d-afe0-1c2256767d4f)
Multiple applications of the Engineering text decal done to test that it
worked as intended.

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Zeta
- add: "Engi" text decal to go with the rest of my decal text set.