Add AI Turrets in Maps (#2043)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Add AI Turrets to Gax, Rad, Meta, Cyberiad, Lambda and Bagel
make minor fixes on Cyberiad

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Gax, Rad, Bagel, Cyberiad, Meta and Lambda now have AI Turrets
- fix: Cyberiad: add missing substation to security, add missing cables
to bridge