Core Rework (#2031)
# Description

Core needs a good amount of touch-ups to EE. I will be updating three
departments and expanding them.

## Engineering Rework:
* Expand Atmos and make a canister room in the SM room for Atmos to get
to SM more easily and to be able to replenish cans much faster.
* We will be changing the SM observation deck and making the SM use
scrubbers instead of passive vents.
* Going to add some more safety features to the SM as well.

### Engie Preview


![image](https://github.com/user-attachments/assets/86795364-accd-4dd8-9899-fc0b72968383)

## Expanding Maintenance 
* Rejoice my grey tiders. I will not have the SM 3 walls thick I will
reduce it to two-ish. Two in some areas one in others. This will be
expanding the maintenance around SM for more loot you can find.

### SM maintenance preview


![image](https://github.com/user-attachments/assets/bef0d0ad-7571-42ab-b673-fc2d22b213ac)

### Old AI core preview

![image](https://github.com/user-attachments/assets/e63f376e-6527-4b84-9121-3a0cf8aa2cf6)

## EPI Rework
* I will be expanding EPI into maintenance to connect the maintenance
room to EPI. I will make a small xenology lab and a can room with this
update.

### Epi rework


![image](https://github.com/user-attachments/assets/2dc211b9-86a2-4b99-987c-bf930761b8a9)

## Add a courtroom
* It seems that NT forgot that we have space law and a courtroom is
needed! That will be added too.

### Courtroom Preview


![image](https://github.com/user-attachments/assets/82225eba-4ef9-4b9c-9ea4-dd0e2f4cf901)


## Update AI core
* Add sentry turrets
* Add AI upload
* Maybe make AI core into a AI sat

### New AI Sat

![image](https://github.com/user-attachments/assets/ab07d956-d8f6-4f40-a860-f328ef179766)

---

# TODO

- [x] Expand Atmosia and fix engineering
- [x] Expand maintenance around the SM core
- [x] Expand EPI to have xenology and to have a Canister room
- [x] Add a courtroom.
- [x] Add an AI sat and move tech vault to not be in maints.
---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added an AI sat, xenobiolab for epi, expanded supermatter maints,
expanded atmos, added a shuttle making dock, and a courtroom.
- tweak: Moved ore silo to vault. Added a spare captain ID locker in
vault.

---------

Co-authored-by: Nathaniel Adams <60526456+Nathaniel-Adams@users.noreply.github.com>